Fix cmake building for DESTDIR
Previously the python binding library was installed inside the build
directory to hand it over to python and strip rpath. This does not work
if DESTDIR is set. The solution is to make a duplicate target of the
pynex library that does not use rpath. The new library without rpath is
handed over to python for installation without the use of cmakes
install.